Refactor git utility files to use `nodegit` namespace
- Renamed `constants.ts`, `createCommit.ts`, `getChanges.ts`, `getStatus.ts`, `getSummaryText.ts`, and `parsePatches.ts` from the `utils/git` directory to the `nodegit` directory.
- Updated import statements in the renamed files to reflect the new file structure.
- Changed import paths for types in the renamed files.
