+++ src/lib/nodegit/constants.ts
