+++ src/lib/nodegit/createCommit.ts
