+++ src/lib/nodegit/getSummaryText.ts
